Review cleanup: bindings parity, CI coverage, integ split, tenet 1 — +parity gate
Closes outstanding REVIEW.md items in one pass.

Bindings gaps
- xa11y-python: CLI entry routes through to_py_err so each xa11y::Error
  variant maps to its typed Python exception instead of collapsing to
  PlatformError.
- xa11y-js: adds Locator.waitUntil(predicate, opts?) (JS-side polling,
  50ms interval, ms timeout) and Subscription.waitFor(predicate, opts?)
  (thin wrapper over waitForEvent('event', …)). Types merged via
  declare module './native.js' { interface Locator { ... } } for the
  pure-JS waitUntil; Subscription.waitFor declared directly. 5 new
  unit tests; 41/41 JS + 137/137 Python unit tests pass.

Windows UIA (tenet 1)
- Previously-flagged let _ = sites (AddProperty, range.Move,
  MoveEndpointByUnit, event Add*Handler registrations) were already
  converted to proper .map_err in prior passes. Remaining let _ =
  sites are legitimate teardown/cleanup paths; added two missing
  one-line WHY comments (uia.rs:1382, 1730).

CI coverage
- New electron job (Linux) runs scripts/run_electron_tests.sh.
- New js-integ-macos job grants TCC kTCCServiceAccessibility to the
  node binary, then runs scripts/run_js_tests.sh under macOS.
- New js-integ-windows job runs the same script via Git Bash
  (shell: bash). Note: the script's `timeout 180 node --test` relies
  on GNU coreutils which Git for Windows on windows-latest ships; if
  it breaks, gate on `command -v timeout`.
- All three gated by `needs: js`. Failure-diagnostics PR-comment
  pattern mirrored for the JS integ jobs.

Integration test split
- xa11y/tests/integ_test.rs shrunk from 2633 → 14 lines.
- Tests split into xa11y/tests/integ/{tree,actions,events_linux,
  events_macos,events_windows}.rs. Helpers stay in integ/mod.rs;
  event modules cfg-gated at the mod declaration. 101 ignored tests
  preserved (65 tree + 25 actions + 11 per-platform events).

Bindings parity gate
- New `cargo xtask check-bindings-parity` parses xa11y-core for
  public methods/fields on App/Locator/Element/Subscription and
  diffs against #[pymethods] blocks + #[pyo3(get)] fields in
  xa11y-python, and against #[napi] impls + #[napi(getter)] methods
  + class/interface declarations in xa11y-js (incl. index.d.ts
  module augmentations). camelCase normalized to snake_case for
  comparison.
- bindings/parity_allowlist.toml covers intentional divergences
  with per-entry rationale: _with provider-injection variants,
  internal plumbing (new/provider/data), Locator::with_timeout
  (bindings use per-call timeouts), Locator::wait_for_state
  (bindings expose specific helpers + predicate-based waitUntil),
  Subscription::{iter,recv,recv_status,try_recv} (bindings use
  idiomatic iterator / EventEmitter patterns).
- Wired into the lint job in ci.yml and into `cargo xtask check`.
