+++ xa11y-windows/src/uia.rs
@@ -1379,6 +1379,7 @@ impl EventContext {
             timestamp: std::time::Instant::now(),
         };
         if let Ok(tx) = self.sender.lock() {
+            // Receiver may be dropped after close(); lost event is expected then.
             let _ = tx.send(event);
         }
     }
@@ -1728,6 +1729,7 @@ impl WindowsProvider {
             )
         } {
             unsafe {
+                // Cleanup during error path; can't override the original error.
                 let _ = self
                     .automation
                     .RemovePropertyChangedEventHandler(&app_root, &property);
