+++ xa11y/tests/integ_test.rs
@@ -6,2628 +6,9 @@
 //! Run with: cargo xtask test-integ
 //!
 //! All tests are `#[ignore]` — the harness script runs them with `--ignored`.
+//!
+//! The tests are split across submodules under `tests/integ/` so each
+//! thematic group lives in its own file while still compiling into this
+//! single `integ_test` binary (keeping compile cost flat).
 
 mod integ;
-
-#[cfg(test)]
-mod tests {
-    use super::integ as h;
-    use xa11y::*;
-
-    // ════════════════════════════════════════════════════════════════
-    // Provider Operations (2 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn check_permissions_granted() {
-        // Permissions are checked automatically by App constructors.
-        // If this fails with PermissionDenied, accessibility or screen
-        // recording permissions are not granted.
-        let _app = h::app_root();
-    }
-
-    #[test]
-    #[ignore]
-    fn apps_returns_nonempty() {
-        let apps = App::list().unwrap();
-        assert!(!apps.is_empty(), "should find at least one application");
-        let has_test_app = apps.iter().any(|a| a.name.contains("xa11y"));
-        assert!(
-            has_test_app,
-            "apps should include the test app. Apps: {:?}",
-            apps.iter().map(|a| &a.name).collect::<Vec<_>>()
-        );
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Tree Structure — Element Discovery (14 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn tree_has_root_node() {
-        let app = h::app_root();
-        assert!(
-            app.data.role == Role::Application || app.data.role == Role::Window,
-            "Root role: {:?}",
-            app.data.role
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_window() {
-        let app = h::app_root();
-        // On Windows (UIA), the app root IS the window — there's no nested
-        // Window child element. Verify root is a Window or find child windows.
-        if app.data.role == Role::Window {
-            return; // App root is the window — pass
-        }
-        let windows = app.locator("window").elements().unwrap();
-        assert!(!windows.is_empty(), "No windows found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_buttons() {
-        let app = h::app_root();
-        let buttons = app.locator("button").elements().unwrap();
-        assert!(
-            buttons.len() >= 2,
-            "Expected >=2 buttons, found {}. App: {}",
-            buttons.len(),
-            app
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_submit_button() {
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        assert_eq!(submit.role, Role::Button);
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_cancel_button_disabled() {
-        let app = h::app_root();
-        let cancel = h::named(&app, "Cancel");
-        // Cancel may have been enabled by a prior toggle test; just verify it exists as a button
-        assert_eq!(cancel.role, Role::Button);
-        // Check that the enabled state is a valid boolean (not that it's a specific value)
-        let _ = cancel.states.enabled;
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_checkbox_unchecked() {
-        let app = h::app_root();
-        let cb = h::named(&app, "I agree to terms");
-        assert_eq!(cb.role, Role::CheckBox);
-        // Checkbox may have been toggled by prior tests; just verify it has a checked state
-        assert!(
-            cb.states.checked.is_some(),
-            "Checkbox should have checked state"
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_text_entry_with_value() {
-        let app = h::app_root();
-        // Prior action tests (TypeText, SetValue) may have changed or cleared the value.
-        // Just verify a text field exists (by role + name), value may or may not be present.
-        let text_elements = app
-            .locator(r#"[role="text_field"]"#)
-            .elements()
-            .unwrap_or_default();
-        let textarea_elements = app
-            .locator(r#"[role="text_area"]"#)
-            .elements()
-            .unwrap_or_default();
-        let has_text = text_elements
-            .iter()
-            .chain(textarea_elements.iter())
-            .any(|n| n.value.is_some() || n.name.as_deref() == Some("Name"));
-        assert!(has_text, "Text entry not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_welcome_label() {
-        let app = h::app_root();
-        // On Linux/AT-SPI with AccessKit, Label nodes may not expose their text
-        // through the Name property or Text interface. Look for the node by name
-        // first, then fall back to checking that StaticText nodes exist.
-        let welcome = app.locator(r#"[name*="Welcome"]"#).elements().unwrap();
-        if welcome.is_empty() {
-            // Fall back: verify that static text nodes exist (labels are present even if unnamed)
-            let labels = app.locator("static_text").elements().unwrap();
-            assert!(
-                !labels.is_empty(),
-                "No StaticText/label nodes found. App: {}",
-                app
-            );
-        } else {
-            assert!(
-                welcome[0].role == Role::StaticText || welcome[0].role == Role::Group,
-                "Welcome node role: {:?}",
-                welcome[0].role
-            );
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_slider_at_50() {
-        let app = h::app_root();
-        let sliders = app.locator("slider").elements().unwrap();
-        assert!(!sliders.is_empty(), "No sliders found. App: {}", app);
-        // Slider value may have been changed by prior tests; just verify it has a numeric value
-        assert!(sliders[0].value.is_some(), "Slider should have a value");
-        let val: f64 = sliders[0].value.as_deref().unwrap().parse().unwrap_or(0.0);
-        assert!(
-            (0.0..=100.0).contains(&val),
-            "Slider value should be in [0,100], got {}",
-            val
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_progress_bar() {
-        let app = h::app_root();
-        let progress = app.locator("progress_bar").elements().unwrap();
-        assert!(!progress.is_empty(), "No progress bars found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_radio_buttons() {
-        let app = h::app_root();
-        let radios = app.locator("radio_button").elements().unwrap();
-        assert!(
-            radios.len() >= 2,
-            "Expected >=2 radio buttons, found {}. App: {}",
-            radios.len(),
-            app
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_combo_box() {
-        let app = h::app_root();
-        let combos = app.locator("combo_box").elements().unwrap();
-        assert!(!combos.is_empty(), "ComboBox not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_list_with_items() {
-        let app = h::app_root();
-        let lists = app.locator("list").elements().unwrap();
-        let items = app.locator("list_item").elements().unwrap();
-        assert!(
-            !lists.is_empty() || !items.is_empty(),
-            "Neither List nor ListItem found. App: {}",
-            app
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_has_table_with_cells() {
-        let app = h::app_root();
-        let tables = app.locator("table").elements().unwrap();
-        let cells = app.locator("table_cell").elements().unwrap();
-        assert!(
-            !tables.is_empty() || !cells.is_empty(),
-            "Neither Table nor TableCell found. App: {}",
-            app
-        );
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Role Coverage (14 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn role_menu_bar() {
-        let app = h::app_root();
-        let nodes = app.locator("menu_bar").elements().unwrap();
-        assert!(!nodes.is_empty(), "MenuBar not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_menu_item() {
-        let app = h::app_root();
-        let nodes = app.locator("menu_item").elements().unwrap();
-        assert!(!nodes.is_empty(), "MenuItem not found. App: {}", app);
-        let has_file = nodes.iter().any(|n| n.name.as_deref() == Some("File"));
-        assert!(has_file, "File menu item not found");
-    }
-
-    #[test]
-    #[ignore]
-    fn role_toolbar() {
-        let app = h::app_root();
-        let nodes = app.locator("toolbar").elements().unwrap();
-        assert!(!nodes.is_empty(), "Toolbar not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_tab_and_tab_group() {
-        let app = h::app_root();
-        let tab_groups = app.locator("tab_group").elements().unwrap();
-        let tabs = app.locator("tab").elements().unwrap();
-        assert!(
-            !tab_groups.is_empty() || !tabs.is_empty(),
-            "Neither TabGroup nor Tab found. App: {}",
-            app
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn role_separator() {
-        let app = h::app_root();
-        let seps = app.locator("separator").elements().unwrap();
-        assert!(!seps.is_empty(), "Separator not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_image() {
-        let app = h::app_root();
-        let images = app.locator("image").elements().unwrap();
-        assert!(!images.is_empty(), "Image not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_link() {
-        let app = h::app_root();
-        let links = app.locator("link").elements().unwrap();
-        assert!(!links.is_empty(), "Link not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_tree_item() {
-        let app = h::app_root();
-        let items = app.locator("tree_item").elements().unwrap();
-        assert!(!items.is_empty(), "TreeItem not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_dialog() {
-        let app = h::app_root();
-        let dialogs = app.locator("dialog").elements().unwrap();
-        assert!(!dialogs.is_empty(), "Dialog not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_alert() {
-        let app = h::app_root();
-        let alerts = app.locator("alert").elements().unwrap();
-        assert!(!alerts.is_empty(), "Alert not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_heading() {
-        let app = h::app_root();
-        let headings = app.locator("heading").elements().unwrap();
-        assert!(!headings.is_empty(), "Heading not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_scroll_bar() {
-        let app = h::app_root();
-        let scrollbars = app.locator("scroll_bar").elements().unwrap();
-        assert!(!scrollbars.is_empty(), "ScrollBar not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_split_group() {
-        let app = h::app_root();
-        // SplitGroup may map through AT-SPI as Group due to accesskit's Pane role
-        let node = app.locator(r#"[name*="SplitGroup"]"#).elements().unwrap();
-        assert!(!node.is_empty(), "SplitGroup node not found. App: {}", app);
-    }
-
-    #[test]
-    #[ignore]
-    fn role_static_text() {
-        let app = h::app_root();
-        let labels = app.locator("static_text").elements().unwrap();
-        assert!(!labels.is_empty(), "StaticText not found. App: {}", app);
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Tree Methods (5 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn tree_children_of_root() {
-        let app = h::app_root();
-        let children = app.children().unwrap();
-        assert!(!children.is_empty(), "Root should have children");
-        // Verify grandchildren have parents (direct children of app root
-        // may report None parent on AT-SPI where the parent is the registry root).
-        if !children.is_empty() {
-            let grandchildren = children[0].children().unwrap();
-            for gc in &grandchildren {
-                let parent = gc.parent().unwrap();
-                assert!(parent.is_some(), "Grandchild should have parent");
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_children_of_leaf() {
-        let app = h::app_root();
-        // Find a leaf node (e.g. a static text or button that has no children)
-        let buttons = app.locator("button").elements().unwrap();
-        for btn in &buttons {
-            let children = btn.children().unwrap();
-            if children.is_empty() {
-                // Found a leaf — verify children returns empty vec, not error
-                assert!(btn.children().unwrap().is_empty());
-                return;
-            }
-        }
-        // If all buttons have children, that's fine too — just verify children() works
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_is_not_empty() {
-        let app = h::app_root();
-        let children = app.children().unwrap();
-        assert!(!children.is_empty(), "Root should have at least one child");
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_display_readable() {
-        let app = h::app_root();
-        let display = app.to_string();
-        assert!(!display.is_empty());
-        // Display should include the app name
-        assert!(
-            display.contains(&app.name),
-            "Display should include app name: {}",
-            display
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn tree_locator_finds_elements() {
-        let app = h::app_root();
-        let buttons = app.locator("button").elements().unwrap();
-        assert!(buttons.len() >= 2, "Expected >=2 buttons via locator");
-        let count = app.locator("button").count().unwrap();
-        assert_eq!(count, buttons.len());
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Element Fields (7 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn element_description_on_image() {
-        let app = h::app_root();
-        let images = app.locator("image").elements().unwrap();
-        if !images.is_empty() {
-            let img = images.iter().find(|n| {
-                n.name.as_deref() == Some("Info Icon")
-                    || n.description.as_deref() == Some("An informational icon")
-            });
-            if let Some(img) = img {
-                assert!(img.description.is_some(), "Image should have description");
-                assert_eq!(img.description.as_deref(), Some("An informational icon"));
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn element_bounds_present() {
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        assert!(submit.bounds.is_some(), "Submit should have bounds");
-        let b = submit.bounds.unwrap();
-        assert!(b.width > 0, "width > 0");
-        assert!(b.height > 0, "height > 0");
-    }
-
-    /// Nodes without the Component interface (e.g. Application root) should
-    /// have `bounds: None` without triggering GTK CRITICAL warnings.
-    #[test]
-    #[ignore]
-    fn element_bounds_none_for_non_component_elements() {
-        let app = h::app_root();
-        // On Linux/macOS, Application elements don't implement Component so
-        // bounds is None. On Windows (UIA), the app root is a Window element
-        // that does have bounds.
-        #[cfg(not(target_os = "windows"))]
-        assert!(
-            app.data.bounds.is_none(),
-            "Application root should not have bounds (no Component interface)"
-        );
-        // But a visible widget like a button should still have bounds
-        let submit = h::named(&app, "Submit");
-        assert!(submit.bounds.is_some(), "Submit button should have bounds");
-    }
-
-    #[test]
-    #[ignore]
-    fn element_actions_list_on_button() {
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        assert!(!submit.actions.is_empty());
-        assert!(
-            submit.actions.iter().any(|a| a == "press"),
-            "Submit should support press, got: {:?}",
-            submit.actions
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn element_children_ids_valid() {
-        let app = h::app_root();
-        let children = app.children().unwrap();
-        assert!(!children.is_empty());
-        for child in &children {
-            // Verify child is a valid element (role may be Unknown for unrecognized elements)
-            let _ = child.role;
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn element_parent_field() {
-        let app = h::app_root();
-        // Direct children of app root may report parent as None on some platforms
-        // (AT-SPI maps parent to registry root which we treat as None).
-        // Test parent on a deeper element instead.
-        let children = app.children().unwrap();
-        if !children.is_empty() {
-            let grandchildren = children[0].children().unwrap();
-            if !grandchildren.is_empty() {
-                let parent = grandchildren[0].parent().unwrap();
-                assert!(parent.is_some(), "Grandchild should have parent");
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn element_handle_nonzero() {
-        let app = h::app_root();
-        // The opaque handle should be non-zero for a valid element
-        assert!(app.data.handle != 0, "Root handle should be nonzero");
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // StateSet Fields (9 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn state_enabled_default() {
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        assert!(submit.states.enabled, "Submit should be enabled");
-    }
-
-    #[test]
-    #[ignore]
-    fn state_disabled_on_cancel() {
-        let app = h::app_root();
-        let cancel = h::named(&app, "Cancel");
-        // Some AT-SPI adapters (AccessKit) may not expose disabled state properly;
-        // in that case, the toggle test (action_toggle_enables_cancel) verifies
-        // the enabled state can change. Here we just verify the node exists and
-        // has a valid enabled state.
-        #[cfg(not(target_os = "linux"))]
-        assert!(!cancel.states.enabled, "Cancel should be disabled");
-        #[cfg(target_os = "linux")]
-        {
-            // On Linux with AccessKit, disabled state may not be reflected.
-            // Just verify the Cancel button exists as a button.
-            assert_eq!(cancel.role, Role::Button);
-            let _ = cancel.states.enabled; // valid boolean either way
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn state_visible_on_shown_widget() {
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        assert!(submit.states.visible, "Submit should be visible");
-    }
-
-    #[test]
-    #[ignore]
-    fn state_focused_after_focus_action() {
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        // Focus action may succeed or fail depending on AT-SPI adapter support
-        let result = h::try_act(&submit, "focus");
-        if result.is_ok() {
-            std::thread::sleep(std::time::Duration::from_millis(100));
-            let app2 = h::app_root();
-            let submit2 = h::named(&app2, "Submit");
-            // Some adapters may not reflect focused state change
-            if !submit2.states.focused {
-                println!("Focus action succeeded but focused state not reflected (AT-SPI adapter limitation)");
-            }
-        } else {
-            println!("Focus action not supported: {:?}", result.err());
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn state_checked_off_on_checkbox() {
-        let app = h::app_root();
-        let cb = h::named(&app, "I agree to terms");
-        assert_eq!(cb.states.checked, Some(Toggled::Off));
-    }
-
-    #[test]
-    #[ignore]
-    fn state_checked_on_radio() {
-        let app = h::app_root();
-        let radios = app.locator("radio_button").elements().unwrap();
-        let opt_a = radios
-            .iter()
-            .find(|n| n.name.as_deref() == Some("Option A"));
-        assert!(opt_a.is_some());
-        assert_eq!(opt_a.unwrap().states.checked, Some(Toggled::On));
-    }
-
-    #[test]
-    #[ignore]
-    fn state_expanded_collapsed_on_expander() {
-        let app = h::app_root();
-        // Look for expandable elements by name
-        let expander_by_name = app.locator(r#"[name*="Expander"]"#).elements().unwrap();
-        // On macOS, GenericContainer with expanded state may not expose AXExpanded.
-        // The expand/collapse actions still work (tested by action_expand_collapse).
-        if expander_by_name.is_empty() {
-            // Verify expand/collapse actions work even if state isn't reported
-            println!(
-                "No expandable elements found by name. \
-                 Expand/collapse actions tested separately."
-            );
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn state_editable_on_text_field() {
-        let app = h::app_root();
-        // Prior action tests (TypeText, SetValue) may have changed or cleared the value.
-        // Find text field by name.
-        let text_fields = app.locator(r#"[name="Name"]"#).elements().unwrap();
-        if text_fields.is_empty() {
-            // Fall back to finding any text field
-            let fields = app.locator("text_field").elements().unwrap();
-            let areas = app.locator("text_area").elements().unwrap();
-            let all_text: Vec<&Element> = fields.iter().chain(areas.iter()).collect();
-            assert!(!all_text.is_empty(), "Text entry not found. App: {}", app);
-            assert!(all_text[0].states.editable, "Text entry should be editable");
-        } else {
-            assert!(
-                text_fields[0].states.editable,
-                "Text entry should be editable"
-            );
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn state_selected_on_list_item() {
-        let app = h::app_root();
-        // Click Apple to select it
-        let apple = h::named(&app, "Apple");
-        let app2 = h::act(&apple, "press");
-        // Verify selection (may come through as Click -> Select depending on AT-SPI mapping)
-        let apple2 = h::named(&app2, "Apple");
-        // Selection might be reported differently; at least verify the action didn't crash
-        println!(
-            "Apple selected state after Click: {:?}",
-            apple2.states.selected
-        );
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Selector Queries (12 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn sel_by_role() {
-        let app = h::app_root();
-        let buttons = app.locator("button").elements().unwrap();
-        assert!(buttons.len() >= 2);
-        for b in &buttons {
-            assert_eq!(b.role, Role::Button);
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_by_exact_name() {
-        let app = h::app_root();
-        let submit = h::one(&app, r#"button[name="Submit"]"#);
-        assert_eq!(submit.role, Role::Button);
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_by_role_and_name() {
-        let app = h::app_root();
-        let results = app.locator(r#"button[name="Cancel"]"#).elements().unwrap();
-        assert_eq!(results.len(), 1);
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_name_contains() {
-        let app = h::app_root();
-        let results = app.locator(r#"[name*="agree"]"#).elements().unwrap();
-        assert!(
-            !results.is_empty(),
-            "Should find element with 'agree' in name"
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_name_starts_with() {
-        let app = h::app_root();
-        // Try "Welc" first (Welcome label), fall back to "Sub" (Submit button)
-        let results = app.locator(r#"[name^="Welc"]"#).elements().unwrap();
-        if results.is_empty() {
-            // Welcome label may not be named on some AT-SPI adapters; use Submit instead
-            let results = app.locator(r#"[name^="Sub"]"#).elements().unwrap();
-            assert!(!results.is_empty());
-            assert!(results[0]
-                .name
-                .as_deref()
-                .unwrap()
-                .to_lowercase()
-                .starts_with("sub"));
-        } else {
-            assert!(results[0]
-                .name
-                .as_deref()
-                .unwrap()
-                .to_lowercase()
-                .starts_with("welc"));
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_name_ends_with() {
-        let app = h::app_root();
-        // "xa11y" suffix may be in the window title or app name
-        let results = app.locator(r#"[name$="xa11y"]"#).elements().unwrap();
-        if results.is_empty() {
-            // Fall back to known name suffixes
-            let results = app.locator(r#"[name$="App"]"#).elements().unwrap();
-            if results.is_empty() {
-                // On Windows, names may differ. Try "Submit" suffix.
-                let results = app.locator(r#"[name$="Submit"]"#).elements().unwrap();
-                assert!(
-                    !results.is_empty(),
-                    "Should find at least one element with name ending in 'Submit'"
-                );
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_value_attribute() {
-        let app = h::app_root();
-        // Try "Red" (ComboBox value), then fall back to any value attribute match.
-        let results = app.locator(r#"[value*="Red"]"#).elements().unwrap();
-        if results.is_empty() {
-            // ComboBox value may not be exposed on some AT-SPI adapters.
-            // Try matching against progress bar value "0.75"
-            let results = app.locator(r#"[value*="0.75"]"#).elements().unwrap();
-            assert!(
-                !results.is_empty(),
-                "Should find element with value containing '0.75' (ProgressBar)"
-            );
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_descendant_combinator() {
-        let app = h::app_root();
-        // On Windows (UIA), the app root IS the window, so "window button"
-        // won't find anything within the app's tree. Use "group button" which
-        // works on all platforms (buttons are inside group containers).
-        let results = app.locator("group button").elements().unwrap();
-        if results.is_empty() {
-            // Fall back to "window button" for Linux/macOS
-            let results = app.locator("window button").elements().unwrap();
-            assert!(!results.is_empty());
-            for r in &results {
-                assert_eq!(r.role, Role::Button);
-            }
-        } else {
-            for r in &results {
-                assert_eq!(r.role, Role::Button);
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_child_combinator() {
-        let app = h::app_root();
-        let results = app.locator("application > window").elements().unwrap();
-        // May or may not match depending on tree structure, but should not error
-        for r in &results {
-            assert_eq!(r.role, Role::Window);
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_nth_pseudo() {
-        let app = h::app_root();
-        let first = app.locator("button:nth(1)").elements().unwrap();
-        assert_eq!(first.len(), 1);
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_role_attribute() {
-        let app = h::app_root();
-        let results = app.locator(r#"[role="button"]"#).elements().unwrap();
-        assert!(!results.is_empty());
-        for r in &results {
-            assert_eq!(r.role, Role::Button);
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn sel_complex_chain() {
-        let app = h::app_root();
-        // Multi-segment selector: role + name attribute chain.
-        // On Windows (UIA), the app root is the window and AccessKit containers
-        // may flatten, so "window button" or "group button" may not work.
-        // Use "menu_bar menu_item" which is nested on all platforms.
-        let results = app
-            .locator(r#"menu_bar menu_item[name="File"]"#)
-            .elements()
-            .unwrap();
-        assert!(!results.is_empty(), "Should find File menu item via chain");
-        assert_eq!(results[0].role, Role::MenuItem);
-        assert_eq!(results[0].name.as_deref(), Some("File"));
-    }
-
-    #[test]
-    #[ignore]
-    fn raw_data_always_present() {
-        let _app = h::app_root();
-        #[cfg(target_os = "linux")]
-        {
-            let atspi_role = _app
-                .data
-                .raw
-                .get("atspi_role")
-                .and_then(|v| v.as_str())
-                .expect("Expected atspi_role in raw data");
-            assert!(!atspi_role.is_empty());
-        }
-        #[cfg(target_os = "macos")]
-        {
-            let ax_role = _app
-                .data
-                .raw
-                .get("ax_role")
-                .and_then(|v| v.as_str())
-                .expect("Expected ax_role in raw data");
-            assert!(!ax_role.is_empty());
-        }
-        #[cfg(target_os = "windows")]
-        {
-            let control_type_id = _app
-                .data
-                .raw
-                .get("control_type_id")
-                .and_then(|v| v.as_i64())
-                .expect("Expected control_type_id in raw data");
-            assert!(control_type_id > 0);
-        }
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Action Dispatch (10 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn action_press_button() {
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        let result = h::try_act(&submit, "press");
-        match result {
-            Ok(()) => println!("Submit pressed"),
-            Err(e) => println!("Submit press result: {}", e),
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn action_toggle_checkbox() {
-        let app = h::app_root();
-        let cbs = app.locator("check_box").elements().unwrap();
-        assert!(!cbs.is_empty(), "No checkbox");
-        let initial = cbs[0].states.checked;
-        let app2 = h::act(&cbs[0], "press");
-        let cb2 = app2.locator("check_box").elements().unwrap();
-        if !cb2.is_empty() {
-            assert_ne!(
-                cb2[0].states.checked, initial,
-                "Checkbox should toggle from {:?}",
-                initial
-            );
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn action_toggle_enables_cancel() {
-        let app = h::app_root();
-        let was_enabled = h::named(&app, "Cancel").states.enabled;
-        let cbs = app.locator("check_box").elements().unwrap();
-        assert!(!cbs.is_empty(), "No checkbox");
-        let app2 = h::act(&cbs[0], "press");
-        let cancel2 = h::named(&app2, "Cancel");
-        // Some AT-SPI adapters may not reflect enabled state changes.
-        // If was_enabled is already true (adapter doesn't report disabled), skip the assertion.
-        if !was_enabled {
-            assert_ne!(cancel2.states.enabled, was_enabled);
-        } else {
-            // Verify the toggle at least didn't crash and Cancel still exists
-            assert_eq!(cancel2.role, Role::Button);
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn action_focus_text_entry() {
-        let app = h::app_root();
-        // Find text entry by name "Name"
-        let text = find_text_entry(&app);
-        let result = h::try_act(&text, "focus");
-        assert!(result.is_ok(), "Focus should succeed: {:?}", result.err());
-    }
-
-    #[test]
-    #[ignore]
-    fn action_set_value_text() {
-        let app = h::app_root();
-        let text = find_text_entry(&app);
-        match text.provider().set_value(&text, "Jane Smith") {
-            Ok(()) => {
-                std::thread::sleep(std::time::Duration::from_millis(300));
-                let app2 = h::app_root();
-                // Value may or may not be reflected via AT-SPI depending on adapter
-                let updated = app2
-                    .locator(r#"[value="Jane Smith"]"#)
-                    .elements()
-                    .unwrap_or_default();
-                if updated.is_empty() {
-                    println!("SetValue succeeded but value not reflected in tree (AT-SPI adapter limitation)");
-                }
-            }
-            Err(Error::TextValueNotSupported) => println!("TextValueNotSupported — OK"),
-            Err(e) => panic!("Unexpected error: {}", e),
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn action_set_value_numeric() {
-        let app = h::app_root();
-        let sliders = app.locator("slider").elements().unwrap();
-        assert!(!sliders.is_empty());
-        let result = sliders[0].provider().set_numeric_value(&sliders[0], 75.0);
-        assert!(result.is_ok(), "SetValue numeric: {:?}", result.err());
-        std::thread::sleep(std::time::Duration::from_millis(300));
-        let app2 = h::app_root();
-        let s2 = app2.locator("slider").elements().unwrap();
-        if !s2.is_empty() {
-            if let Some(v) = &s2[0].value {
-                let val: f64 = v.parse().unwrap_or(0.0);
-                assert!(
-                    (val - 75.0).abs() < 2.0,
-                    "Slider should be ~75, got {}",
-                    val
-                );
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn action_increment_spinner() {
-        let app = h::app_root();
-        // Find spin button or slider with a numeric value
-        let sliders = app.locator("slider").elements().unwrap();
-        let spin = sliders.first();
-        if let Some(spin) = spin {
-            let initial: f64 = spin.value.as_deref().unwrap_or("0").parse().unwrap_or(0.0);
-            let result = h::try_act(spin, "increment");
-            if result.is_ok() {
-                std::thread::sleep(std::time::Duration::from_millis(300));
-                let app2 = h::app_root();
-                if let Some(s2) = app2.locator("slider").elements().unwrap().first() {
-                    if let Some(v) = &s2.value {
-                        let new_val: f64 = v.parse().unwrap_or(initial);
-                        assert!(
-                            new_val > initial,
-                            "Value should increase: {} -> {}",
-                            initial,
-                            new_val
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn action_decrement_spinner() {
-        let app = h::app_root();
-        let sliders = app.locator("slider").elements().unwrap();
-        let spin = sliders.first();
-        if let Some(spin) = spin {
-            let before: f64 = spin.value.as_deref().unwrap_or("0").parse().unwrap_or(0.0);
-            let result = h::try_act(spin, "decrement");
-            if result.is_ok() {
-                std::thread::sleep(std::time::Duration::from_millis(300));
-                let app2 = h::app_root();
-                if let Some(s2) = app2.locator("slider").elements().unwrap().first() {
-                    if let Some(v) = &s2.value {
-                        let after: f64 = v.parse().unwrap_or(before);
-                        assert!(
-                            after < before,
-                            "Value should decrease: {} -> {}",
-                            before,
-                            after
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn action_expand_collapse() {
-        let app = h::app_root();
-        let expander = app
-            .locator(r#"[name*="Expander"]"#)
-            .elements()
-            .unwrap()
-            .into_iter()
-            .next();
-        if let Some(node) = expander {
-            // Expand
-            if h::try_act(&node, "expand").is_ok() {
-                std::thread::sleep(std::time::Duration::from_millis(300));
-                let app2 = h::app_root();
-                let n2 = app2
-                    .locator(r#"[name*="Expander"]"#)
-                    .elements()
-                    .unwrap()
-                    .into_iter()
-                    .next();
-                if let Some(n) = n2 {
-                    if n.states.expanded == Some(true) {
-                        // Collapse
-                        if h::try_act(&n, "collapse").is_ok() {
-                            std::thread::sleep(std::time::Duration::from_millis(300));
-                            let app3 = h::app_root();
-                            let n3 = app3
-                                .locator(r#"[name*="Expander"]"#)
-                                .elements()
-                                .unwrap()
-                                .into_iter()
-                                .next();
-                            if let Some(n) = n3 {
-                                assert_eq!(n.states.expanded, Some(false));
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn action_select_list_item() {
-        let app = h::app_root();
-        let apple = app.locator(r#"[name*="Apple"]"#).elements().unwrap();
-        if !apple.is_empty() {
-            let _ = h::try_act(&apple[0], "press");
-            // Selection verified by not crashing; state_selected_on_list_item tests the state
-        }
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Complex / Stress Scenarios (5 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn nesting_deep_tree_traversal() {
-        let app = h::app_root();
-        // Query inside table -> row -> cell
-        let cells = app.locator(r#"[name*="Alice"]"#).elements().unwrap();
-        assert!(!cells.is_empty(), "Alice cell not found. App: {}", app);
-        // Verify nesting: cell's parent should exist
-        let parent = cells[0].parent().unwrap();
-        assert!(parent.is_some());
-    }
-
-    #[test]
-    #[ignore]
-    fn nesting_subtree_of_table() {
-        let app = h::app_root();
-        let tables = app.locator("table").elements().unwrap();
-        if !tables.is_empty() {
-            // Table should contain rows and cells — use descendant combinator.
-            // On Windows (UIA), AccessKit Cell maps to DataItem (table_row) since
-            // UIA has no distinct Cell control type for data grids.
-            let cells = app.locator("table table_cell").elements().unwrap();
-            if cells.len() >= 2 {
-                return;
-            }
-            // Fall back to table_row children (Windows: cells appear as table_row)
-            let rows = app.locator("table table_row").elements().unwrap();
-            assert!(
-                rows.len() >= 2,
-                "Table should have at least 2 rows/cells, found {}. App: {}",
-                rows.len(),
-                app
-            );
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn thrash_toggle_checkbox_5_times() {
-        let app = h::app_root();
-        let cbs = app.locator("check_box").elements().unwrap();
-        assert!(!cbs.is_empty());
-        let mut current_app = app;
-        for _ in 0..5 {
-            let cbs = current_app.locator("check_box").elements().unwrap();
-            assert!(!cbs.is_empty());
-            current_app = h::act(&cbs[0], "press");
-        }
-        // After 5 toggles (odd), state should have flipped from initial
-        let final_cb = current_app.locator("check_box").elements().unwrap();
-        if !final_cb.is_empty() {
-            assert_eq!(
-                final_cb[0].states.checked,
-                Some(Toggled::On),
-                "After 5 toggles from Off, should be On"
-            );
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn thrash_slider_increment_10_times() {
-        let app = h::app_root();
-        let sliders = app.locator("slider").elements().unwrap();
-        let slider = sliders.first().expect("No slider");
-        let start_val: f64 = slider
-            .value
-            .as_deref()
-            .unwrap_or("0")
-            .parse()
-            .unwrap_or(0.0);
-        let mut current_app = app;
-        for _ in 0..10 {
-            let sliders = current_app.locator("slider").elements().unwrap();
-            let slider = sliders.first().expect("No slider");
-            current_app = h::act(slider, "increment");
-        }
-        let s = current_app.locator("slider").elements().unwrap();
-        if !s.is_empty() {
-            if let Some(v) = &s[0].value {
-                let val: f64 = v.parse().unwrap_or(0.0);
-                let expected = (start_val + 10.0).min(100.0);
-                assert!(
-                    (val - expected).abs() < 2.0,
-                    "After 10 increments from {}, should be ~{}, got {}",
-                    start_val,
-                    expected,
-                    val
-                );
-            }
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn thrash_expand_collapse_cycle() {
-        let app = h::app_root();
-        let has_expander = !app
-            .locator(r#"[name*="Expander"]"#)
-            .elements()
-            .unwrap()
-            .is_empty();
-        if has_expander {
-            let mut ct = app;
-            // expand, collapse, expand, collapse
-            for action in ["expand", "collapse", "expand", "collapse"] {
-                let node = ct
-                    .locator(r#"[name*="Expander"]"#)
-                    .elements()
-                    .unwrap()
-                    .into_iter()
-                    .next()
-                    .expect("Expander node should exist");
-                ct = h::act(&node, action);
-            }
-            let final_node = ct
-                .locator(r#"[name*="Expander"]"#)
-                .elements()
-                .unwrap()
-                .into_iter()
-                .next();
-            if let Some(n) = final_node {
-                if n.states.expanded.is_some() {
-                    assert_eq!(n.states.expanded, Some(false), "Should end collapsed");
-                }
-            }
-        }
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Error Paths (4 tests)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn error_app_not_found() {
-        let result = App::by_name("nonexistent_app_12345");
-        assert!(result.is_err());
-        assert!(matches!(
-            result.unwrap_err(),
-            Error::SelectorNotMatched { .. }
-        ));
-    }
-
-    #[test]
-    #[ignore]
-    fn error_selector_not_matched() {
-        let app = h::app_root();
-        let result = app
-            .locator(r#"button[name="nonexistent_element_12345"]"#)
-            .elements();
-        assert!(result.unwrap().is_empty());
-    }
-
-    #[test]
-    #[ignore]
-    fn error_invalid_selector() {
-        let app = h::app_root();
-        let result = app.locator("$$$invalid!!!").elements();
-        assert!(result.is_err());
-        assert!(matches!(result.unwrap_err(), Error::InvalidSelector { .. }));
-    }
-
-    #[test]
-    #[ignore]
-    fn action_on_default_tree() {
-        let app = h::app_root();
-        let buttons = app.locator(r#"[name*="Submit"]"#).elements().unwrap();
-        assert!(!buttons.is_empty());
-        let result = h::try_act(&buttons[0], "press");
-        match result {
-            Ok(()) => {}
-            Err(e) => assert!(
-                matches!(e, Error::Platform { .. } | Error::ElementStale { .. }),
-                "Unexpected error: {}",
-                e
-            ),
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn error_press_on_non_interactive_element() {
-        // Pressing a static_text or other non-interactive element should fail.
-        // On macOS this returns ActionNotSupported; on Linux it returns Platform.
-        let app = h::app_root();
-        let labels = app.locator("static_text").elements().unwrap_or_default();
-        if let Some(label) = labels
-            .into_iter()
-            .find(|e| !e.data().actions.iter().any(|a| a == "press"))
-        {
-            let result = h::try_act(&label, "press");
-            assert!(
-                result.is_err(),
-                "Press on static_text should fail: {:?}",
-                result
-            );
-            #[cfg(target_os = "macos")]
-            assert!(
-                matches!(result, Err(Error::ActionNotSupported { .. })),
-                "Expected ActionNotSupported on macOS, got: {:?}",
-                result
-            );
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn error_expand_on_non_expandable_element() {
-        // Expanding a button (not expandable) should return an error,
-        // not silently succeed.
-        let app = h::app_root();
-        let button = h::named(&app, "Submit");
-        let result = h::try_act(&button, "expand");
-        assert!(
-            result.is_err(),
-            "Expand on a non-expandable button should fail, not silently succeed"
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn error_collapse_on_non_expandable_element() {
-        // Collapsing a button (not expandable) should return an error,
-        // not silently succeed.
-        let app = h::app_root();
-        let button = h::named(&app, "Submit");
-        let result = h::try_act(&button, "collapse");
-        assert!(
-            result.is_err(),
-            "Collapse on a non-expandable button should fail, not silently succeed"
-        );
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Serialization (1 test)
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn json_roundtrip_real_element() {
-        let app = h::app_root();
-        // Serialize the root ElementData
-        let json = serde_json::to_string(&app.data).unwrap();
-        let deser: ElementData = serde_json::from_str(&json).unwrap();
-        assert_eq!(deser.role, app.data.role);
-        assert_eq!(deser.name, app.data.name);
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // New Actions — Blur, Scroll, SetTextSelection, TypeText
-    // ════════════════════════════════════════════════════════════════
-
-    #[test]
-    #[ignore]
-    fn action_blur_text_entry() {
-        let app = h::app_root();
-        let text = find_text_entry(&app);
-
-        // Focus first
-        let result = h::try_act(&text, "focus");
-        assert!(result.is_ok(), "Focus should succeed: {:?}", result.err());
-
-        // Then blur — re-find the text entry from a fresh root
-        let app2 = h::app_root();
-        let text2 = find_text_entry(&app2);
-        let result = h::try_act(&text2, "blur");
-        assert!(result.is_ok(), "Blur should succeed: {:?}", result.err());
-    }
-
-    #[test]
-    #[ignore]
-    fn action_set_text_selection() {
-        let app = h::app_root();
-        let text = find_text_entry(&app);
-
-        // Focus first
-        let _ = h::try_act(&text, "focus");
-
-        // Select characters 0..4 ("John")
-        let result = text.provider().set_text_selection(&text, 0, 4);
-        match result {
-            Ok(()) => println!("SetTextSelection succeeded"),
-            Err(e) => println!("SetTextSelection result: {}", e),
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn action_type_text() {
-        let app = h::app_root();
-        let text = find_text_entry(&app);
-
-        // Focus first
-        let _ = h::try_act(&text, "focus");
-
-        // Type text
-        let result = text.provider().type_text(&text, "hi");
-        match result {
-            Ok(()) => println!("TypeText succeeded"),
-            Err(e) => println!("TypeText result: {}", e),
-        }
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Event subscription (macOS-only end-to-end tests)
-    // ════════════════════════════════════════════════════════════════
-    //
-    // These tests exercise the native event subscription API against the
-    // AccessKit test app on macOS. Linux and Windows providers currently
-    // ship with inert subscription stubs (no events are ever delivered)
-    // per the events design doc — their backends will be implemented
-    // separately.
-    //
-    // AccessKit's macOS bridge (accesskit_macos::EventGenerator) reliably
-    // emits the following notifications, which the xa11y-macos provider
-    // maps to EventKind variants:
-    //
-    //     Cocoa notification             → xa11y EventKind
-    //     AXFocusedUIElementChanged      → FocusChanged
-    //     AXValueChanged                 → ValueChanged (+ TextChanged for
-    //                                       text roles, + StateChanged{Checked}
-    //                                       for checkbox/radio roles)
-    //     AXTitleChanged                 → NameChanged
-    //     AXUIElementDestroyed           → StructureChanged
-    //     AXSelectedTextChanged          → SelectionChanged
-    //     AXAnnouncementRequested        → Announcement (live-region updates)
-    //
-    // AccessKit's macOS bridge does NOT emit the following — they require
-    // native NSMenu/NSWindow behavior that the AccessKit adapter does not
-    // synthesize. The macOS provider still subscribes to them, so they
-    // will propagate correctly when a non-AccessKit app raises them, but
-    // the AccessKit test app cannot drive e2e coverage for them today:
-    //
-    //     MenuOpened / MenuClosed              (NSMenu only)
-    //     WindowOpened / WindowClosed          (multi-window required)
-    //     WindowActivated / WindowDeactivated  (key-window change required)
-    //     StateChanged { Busy } (and other flags not backed by value-changes)
-    //
-    // IMPORTANT: these tests MUST fail on timeout. A previous iteration
-    // caught `Error::Timeout` and logged a "may depend on platform"
-    // message while reporting success, hiding real regressions. Do not
-    // reintroduce that pattern.
-
-    #[cfg(target_os = "macos")]
-    fn find_name_field(app: &App) -> Element {
-        app.locator(r#"[name="Name"]"#)
-            .elements()
-            .unwrap_or_default()
-            .into_iter()
-            .next()
-            .expect("Name text field not found in test app")
-    }
-
-    #[cfg(target_os = "macos")]
-    fn ensure_checkbox(app: &App, want_on: bool) {
-        let chk = app
-            .locator("check_box")
-            .element()
-            .expect("check_box not found");
-        let is_on = chk.states.checked == Some(Toggled::On);
-        if is_on != want_on {
-            chk.provider().toggle(&chk).expect("toggle failed");
-            std::thread::sleep(std::time::Duration::from_millis(150));
-        }
-    }
-
-    // ── Subscription mechanics ──
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_try_recv_returns_none_when_idle() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-
-        // Drain anything that may have been queued during subscription setup.
-        std::thread::sleep(Duration::from_millis(100));
-        while sub.try_recv().is_some() {}
-
-        // After draining, try_recv must return None without blocking.
-        assert!(sub.try_recv().is_none());
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_recv_times_out_when_idle() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-
-        std::thread::sleep(Duration::from_millis(100));
-        while sub.try_recv().is_some() {}
-
-        let r = sub.recv(Duration::from_millis(300));
-        assert!(
-            matches!(r, Err(Error::Timeout { .. })),
-            "expected Timeout, got {:?}",
-            r
-        );
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_drop_unsubscribes_cleanly() {
-        let app = h::app_root();
-        {
-            let _sub = app.subscribe().expect("subscribe");
-        }
-        // Re-subscribing after drop must not hang or fail.
-        let _sub2 = app.subscribe().expect("re-subscribe");
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_metadata_populated() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let expected_pid = app.pid;
-
-        // Use the slider to deterministically drive an event — setting a
-        // value different from its current one always fires AXValueChanged
-        // (no dependence on prior test state).
-        let slider = app.locator(r#"[role="slider"]"#).element().expect("slider");
-        let target_val = if slider.numeric_value == Some(42.0) {
-            17.0
-        } else {
-            42.0
-        };
-
-        let sub = app.subscribe().expect("subscribe");
-        slider
-            .provider()
-            .set_numeric_value(&slider, target_val)
-            .expect("set_numeric_value");
-
-        let event = sub
-            .wait_for(|_| true, Duration::from_secs(3))
-            .expect("at least one event must arrive within 3s");
-
-        if let Some(pid) = expected_pid {
-            assert_eq!(event.app_pid, pid, "app_pid must match subscribed app");
-        }
-        assert!(!event.app_name.is_empty(), "app_name must be populated");
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_recv_delivers_across_threads() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-
-        let handle =
-            std::thread::spawn(move || -> Result<Event> { sub.recv(Duration::from_secs(5)) });
-
-        // Let the thread block on recv before we trigger anything.
-        std::thread::sleep(Duration::from_millis(100));
-
-        let btn = h::named(&h::app_root(), "Submit");
-        btn.provider().focus(&btn).expect("focus failed");
-
-        let event = handle
-            .join()
-            .expect("recv thread panicked")
-            .expect("recv must yield an event");
-        assert!(!event.app_name.is_empty());
-    }
-
-    // ── Per-EventKind end-to-end tests ──
-
-    /// FocusChanged: focus a button other than the currently-focused one.
-    /// AccessKit fires AXFocusedUIElementChanged on every focus move.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_focus_changed() {
-        use std::time::Duration;
-        let app = h::app_root();
-        // Focus defaults to Submit. Focus Cancel to force a change.
-        let target = app
-            .locator(r#"button[name="Cancel"]"#)
-            .element()
-            .expect("Cancel button not found");
-
-        let sub = app.subscribe().expect("subscribe");
-        target.provider().focus(&target).expect("focus failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::FocusChanged,
-                Duration::from_secs(3),
-            )
-            .expect("FocusChanged must be delivered within 3s");
-
-        assert_eq!(event.kind, EventKind::FocusChanged);
-        let tgt = event.target.expect("FocusChanged target must be populated");
-        assert_eq!(
-            tgt.role,
-            Role::Button,
-            "expected Button, got {:?}",
-            tgt.role
-        );
-    }
-
-    /// ValueChanged: set a slider to a new value. AccessKit fires
-    /// AXValueChanged whenever node.raw_value() changes.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_value_changed() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let slider = app
-            .locator(r#"[role="slider"]"#)
-            .element()
-            .expect("slider not found");
-
-        let sub = app.subscribe().expect("subscribe");
-        slider
-            .provider()
-            .set_numeric_value(&slider, 73.0)
-            .expect("set_numeric_value failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::ValueChanged,
-                Duration::from_secs(3),
-            )
-            .expect("ValueChanged must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::ValueChanged);
-        let tgt = event.target.expect("target");
-        assert_eq!(tgt.role, Role::Slider);
-    }
-
-    /// NameChanged: press Submit, which changes the status label text from
-    /// "Status: Ready" to "Status: Please agree to terms" (checkbox off) or
-    /// "Status: Submitted" (checkbox on). AccessKit fires AXTitleChanged.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_name_changed() {
-        use std::time::Duration;
-
-        // Step 1: prime status to "Please agree to terms" (checkbox off).
-        let app = h::app_root();
-        ensure_checkbox(&app, false);
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        submit
-            .provider()
-            .press(&submit)
-            .expect("prime press failed");
-        std::thread::sleep(Duration::from_millis(200));
-
-        // Step 2: flip checkbox on so the next Submit press drives status
-        // to "Submitted", guaranteed distinct from step 1.
-        let app = h::app_root();
-        ensure_checkbox(&app, true);
-        std::thread::sleep(Duration::from_millis(200));
-
-        // Step 3: subscribe, press Submit, assert NameChanged.
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-        let submit = h::named(&app, "Submit");
-        submit
-            .provider()
-            .press(&submit)
-            .expect("trigger press failed");
-
-        let event = sub
-            .wait_for(|e| e.kind == EventKind::NameChanged, Duration::from_secs(3))
-            .expect("NameChanged must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::NameChanged);
-    }
-
-    /// TextChanged: set a text field's value. AccessKit fires AXValueChanged
-    /// on a text role; the macOS backend also synthesizes TextChanged.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_text_changed() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let text = find_name_field(&app);
-
-        let sub = app.subscribe().expect("subscribe");
-        text.provider()
-            .set_value(&text, "Event E2E Text")
-            .expect("set_value failed");
-
-        let event = sub
-            .wait_for(|e| e.kind == EventKind::TextChanged, Duration::from_secs(3))
-            .expect("TextChanged must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::TextChanged);
-        let tgt = event.target.expect("target");
-        assert!(
-            matches!(tgt.role, Role::TextField | Role::TextArea),
-            "expected text role, got {:?}",
-            tgt.role
-        );
-    }
-
-    /// StateChanged { Checked }: toggling the checkbox fires AXValueChanged
-    /// on a role of AXCheckBox; the macOS backend also emits StateChanged
-    /// with the new checked flag.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_state_changed_checked() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let chk = app
-            .locator("check_box")
-            .element()
-            .expect("check_box not found");
-        let was_on = chk.states.checked == Some(Toggled::On);
-
-        let sub = app.subscribe().expect("subscribe");
-        chk.provider().toggle(&chk).expect("toggle failed");
-
-        let event = sub
-            .wait_for(
-                |e| {
-                    matches!(
-                        e.kind,
-                        EventKind::StateChanged {
-                            flag: StateFlag::Checked,
-                            ..
-                        }
-                    )
-                },
-                Duration::from_secs(3),
-            )
-            .expect("StateChanged{Checked} must be delivered within 3s");
-
-        match event.kind {
-            EventKind::StateChanged {
-                flag: StateFlag::Checked,
-                value,
-            } => {
-                assert_eq!(value, !was_on, "checked flag must flip");
-            }
-            other => panic!("unexpected kind: {:?}", other),
-        }
-    }
-
-    // StructureChanged: no end-to-end test yet.
-    //
-    // AccessKit's macOS bridge posts AXUIElementDestroyedNotification on the
-    // element that is being torn down, but macOS does NOT propagate that
-    // specific notification to observers registered on ancestor elements —
-    // empirically confirmed by enabling the registration trace in our
-    // subscribe path: all notifications register successfully, and other
-    // kinds (focus, value, title, announcement) propagate to the app-level
-    // observer as expected, but AXUIElementDestroyed fired by AccessKit's
-    // subtree removal never reaches it.
-    //
-    // The provider is set up correctly to dispatch StructureChanged when
-    // AXUIElementDestroyed does reach the callback (see
-    // ax_observer_callback). Driving it requires either:
-    //   - per-element observer registration (complex — requires tracking
-    //     element lifetime), or
-    //   - a non-AccessKit test app (Cocoa/AppKit) where NSWindow/NSView
-    //     teardown fires the notification on the application element.
-    //
-    // Follow-up: add a Cocoa integration test harness or element-scoped
-    // subscription support and cover StructureChanged there.
-
-    // SelectionChanged: no end-to-end test yet.
-    //
-    // - AXSelectedTextChangedNotification requires `supports_text_ranges` on
-    //   the text node, which means AccessKit text-run children. The test
-    //   app's TextInput doesn't model runs, so set_text_selection lands at
-    //   the AX level but AccessKit never synthesizes the notification.
-    // - AXSelectedRowsChangedNotification / AXSelectedChildrenChanged are
-    //   element-scoped on macOS; our app-level observer doesn't receive them
-    //   for descendants by default, so driving via list-item selection
-    //   doesn't surface them either.
-    //
-    // Follow-up: extend the AccessKit test app with text-run-backed
-    // selectable text, or add per-container observer registration for
-    // selection notifications.
-
-    /// Announcement: press the test app's "Announce" button, which updates
-    /// a Live::Polite region's value. AccessKit's macOS bridge fires
-    /// AXAnnouncementRequested on the owning window.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "macos")]
-    fn event_announcement() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let btn = app.locator(r#"button[name="Announce"]"#).element().expect(
-            "Announce button not found in test app — \
-                 required for Announcement e2e test",
-        );
-
-        let sub = app.subscribe().expect("subscribe");
-        btn.provider().press(&btn).expect("Announce press failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::Announcement,
-                Duration::from_secs(3),
-            )
-            .expect("Announcement must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::Announcement);
-    }
-
-    // ════════════════════════════════════════════════════════════════
-    // Event subscription (Windows-only end-to-end tests)
-    // ════════════════════════════════════════════════════════════════
-    //
-    // These tests exercise the native UIA event subscription backend
-    // against the AccessKit test app. They are `#[ignore]` like all the
-    // other integration tests. CI runs them on windows-latest via
-    // `scripts/run_integ_tests_windows.ps1`; the same runner class already
-    // executes the Qt UIA tests, so the "no interactive desktop" objection
-    // that earlier versions of this doc raised is factually wrong.
-    //
-    // AccessKit's Windows bridge (accesskit_windows) emits UIA events via
-    // `UiaRaiseAutomationEvent`, `UiaRaiseAutomationPropertyChangedEvent`,
-    // and `UiaRaiseStructureChangedEvent` whenever a tree update touches
-    // the relevant properties:
-    //
-    //     UIA notification                     → xa11y EventKind
-    //     AddFocusChangedEventHandler          → FocusChanged
-    //     PropertyChanged(Value_Value)         → ValueChanged
-    //     PropertyChanged(RangeValue_Value)    → ValueChanged
-    //     PropertyChanged(ToggleState)         → ValueChanged + StateChanged{Checked}
-    //     PropertyChanged(Name)                → NameChanged
-    //     PropertyChanged(IsEnabled)           → StateChanged{Enabled}
-    //     PropertyChanged(ExpandCollapseState) → StateChanged{Expanded}
-    //     UIA_Text_TextChangedEventId          → TextChanged
-    //     UIA_NotificationEventId              → Announcement
-    //     UIA_LiveRegionChangedEventId         → Announcement
-    //     UIA_Window_WindowOpenedEventId       → WindowOpened
-    //     StructureChangedEventHandler         → StructureChanged
-    //
-    // AccessKit's Windows bridge does NOT synthesize certain events that
-    // only native Win32/WPF controls would raise — notably the menu
-    // open/close pair (UIA_MenuOpenedEventId / UIA_MenuClosedEventId).
-    // Our provider still registers them so non-AccessKit apps get them.
-    //
-    // Unlike the macOS tests, these tests MUST NOT catch Error::Timeout
-    // and pass silently — a hard panic on timeout is the only way to
-    // surface real regressions.
-
-    #[cfg(target_os = "windows")]
-    fn find_name_field_win(app: &App) -> Element {
-        app.locator(r#"[name="Name"]"#)
-            .elements()
-            .unwrap_or_default()
-            .into_iter()
-            .next()
-            .expect("Name text field not found in test app")
-    }
-
-    #[cfg(target_os = "windows")]
-    fn ensure_checkbox_win(app: &App, want_on: bool) {
-        let chk = app
-            .locator("check_box")
-            .element()
-            .expect("check_box not found");
-        let is_on = chk.states.checked == Some(Toggled::On);
-        if is_on != want_on {
-            chk.provider().toggle(&chk).expect("toggle failed");
-            std::thread::sleep(std::time::Duration::from_millis(150));
-        }
-    }
-
-    // ── Subscription mechanics ──
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_try_recv_returns_none_when_idle_win() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-
-        // Drain anything from subscription setup (UIA often replays focus).
-        std::thread::sleep(Duration::from_millis(200));
-        while sub.try_recv().is_some() {}
-
-        assert!(sub.try_recv().is_none());
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_recv_times_out_when_idle_win() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-
-        std::thread::sleep(Duration::from_millis(200));
-        while sub.try_recv().is_some() {}
-
-        let r = sub.recv(Duration::from_millis(300));
-        assert!(
-            matches!(r, Err(Error::Timeout { .. })),
-            "expected Timeout, got {:?}",
-            r
-        );
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_drop_unsubscribes_cleanly_win() {
-        let app = h::app_root();
-        {
-            let _sub = app.subscribe().expect("subscribe");
-        }
-        // Re-subscribing after drop must not hang or fail: RemoveXxx runs
-        // synchronously and per-handler, so the second subscribe starts
-        // with a clean slate.
-        let _sub2 = app.subscribe().expect("re-subscribe");
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_metadata_populated_win() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let expected_pid = app.pid;
-
-        // Drive a deterministic event via the slider so we don't depend on
-        // prior test state — setting a distinct value always fires
-        // PropertyChanged(RangeValue_Value).
-        let slider = app.locator(r#"[role="slider"]"#).element().expect("slider");
-        let target_val = if slider.numeric_value == Some(42.0) {
-            17.0
-        } else {
-            42.0
-        };
-
-        let sub = app.subscribe().expect("subscribe");
-        slider
-            .provider()
-            .set_numeric_value(&slider, target_val)
-            .expect("set_numeric_value");
-
-        let event = sub
-            .wait_for(|_| true, Duration::from_secs(3))
-            .expect("at least one event must arrive within 3s");
-
-        if let Some(pid) = expected_pid {
-            assert_eq!(event.app_pid, pid, "app_pid must match subscribed app");
-        }
-        assert!(!event.app_name.is_empty(), "app_name must be populated");
-    }
-
-    // ── Per-EventKind end-to-end tests ──
-
-    /// FocusChanged: focus a non-default button to force a focus move.
-    /// Targets the always-enabled "New" toolbar button rather than
-    /// "Cancel" — Cancel starts disabled, and the Windows UIA SetFocus
-    /// call fails (`E_INVALIDARG`, HRESULT 0x80070057 when the provider
-    /// reports IsEnabled=false) against disabled elements. The earlier
-    /// test ran only locally (via `run_integ_tests_windows.ps1`, where
-    /// prior tests may have enabled Cancel by toggling the checkbox);
-    /// now that CI runs the suite cold it needs a target that's
-    /// guaranteed focusable from a fresh app launch.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_focus_changed_win() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let target = app
-            .locator(r#"button[name="New"]"#)
-            .element()
-            .expect("New toolbar button not found");
-
-        let sub = app.subscribe().expect("subscribe");
-        target.provider().focus(&target).expect("focus failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::FocusChanged,
-                Duration::from_secs(3),
-            )
-            .expect("FocusChanged must be delivered within 3s");
-
-        assert_eq!(event.kind, EventKind::FocusChanged);
-        let tgt = event.target.expect("FocusChanged target must be populated");
-        assert_eq!(
-            tgt.role,
-            Role::Button,
-            "expected Button, got {:?}",
-            tgt.role
-        );
-    }
-
-    /// ValueChanged: setting the slider fires PropertyChanged(RangeValue_Value).
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_value_changed_win() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let slider = app
-            .locator(r#"[role="slider"]"#)
-            .element()
-            .expect("slider not found");
-
-        let sub = app.subscribe().expect("subscribe");
-        slider
-            .provider()
-            .set_numeric_value(&slider, 73.0)
-            .expect("set_numeric_value failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::ValueChanged,
-                Duration::from_secs(3),
-            )
-            .expect("ValueChanged must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::ValueChanged);
-        let tgt = event.target.expect("target");
-        assert_eq!(tgt.role, Role::Slider);
-    }
-
-    /// NameChanged: pressing Submit mutates the status label's name,
-    /// which AccessKit translates into PropertyChanged(Name).
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_name_changed_win() {
-        use std::time::Duration;
-
-        // Step 1: prime status to "Please agree to terms" (checkbox off).
-        let app = h::app_root();
-        ensure_checkbox_win(&app, false);
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        submit
-            .provider()
-            .press(&submit)
-            .expect("prime press failed");
-        std::thread::sleep(Duration::from_millis(200));
-
-        // Step 2: flip checkbox on so the next Submit press drives status
-        // to "Submitted", guaranteed distinct from step 1.
-        let app = h::app_root();
-        ensure_checkbox_win(&app, true);
-        std::thread::sleep(Duration::from_millis(200));
-
-        // Step 3: subscribe, press Submit, assert NameChanged.
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-        let submit = h::named(&app, "Submit");
-        submit
-            .provider()
-            .press(&submit)
-            .expect("trigger press failed");
-
-        let event = sub
-            .wait_for(|e| e.kind == EventKind::NameChanged, Duration::from_secs(3))
-            .expect("NameChanged must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::NameChanged);
-    }
-
-    /// StateChanged{Checked}: toggling the checkbox fires
-    /// PropertyChanged(ToggleState) with the new boolean value.
-    /// The Windows backend also emits ValueChanged alongside — this test
-    /// only asserts the StateChanged variant is delivered.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_state_changed_checked_win() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let chk = app
-            .locator("check_box")
-            .element()
-            .expect("check_box not found");
-        let was_on = chk.states.checked == Some(Toggled::On);
-
-        let sub = app.subscribe().expect("subscribe");
-        chk.provider().toggle(&chk).expect("toggle failed");
-
-        let event = sub
-            .wait_for(
-                |e| {
-                    matches!(
-                        e.kind,
-                        EventKind::StateChanged {
-                            flag: StateFlag::Checked,
-                            ..
-                        }
-                    )
-                },
-                Duration::from_secs(3),
-            )
-            .expect("StateChanged{Checked} must be delivered within 3s");
-
-        match event.kind {
-            EventKind::StateChanged {
-                flag: StateFlag::Checked,
-                value,
-            } => {
-                assert_eq!(value, !was_on, "checked flag must flip");
-            }
-            other => panic!("unexpected kind: {:?}", other),
-        }
-
-        // Restore the checkbox to its pre-test state so later tests in the
-        // same `cargo test --test-threads=1` run (state_checked_off_on_
-        // checkbox, state_disabled_on_cancel, thrash_toggle_checkbox_5_
-        // times) see the app in its expected initial state. Toggling the
-        // checkbox also drives `state.cancel_enabled`, so a dirty checkbox
-        // leaks into multiple fixtures. Matches the Linux test's cleanup.
-        let app = h::app_root();
-        let chk = app
-            .locator("check_box")
-            .element()
-            .expect("check_box not found (post-test)");
-        if (chk.states.checked == Some(Toggled::On)) != was_on {
-            chk.provider()
-                .toggle(&chk)
-                .expect("post-test toggle to restore checkbox");
-        }
-    }
-
-    /// ToggleState also produces a ValueChanged event (the design doc
-    /// says the ToggleState path emits both). This test asserts the
-    /// ValueChanged half of the same interaction.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_toggle_emits_value_changed_win() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let chk = app
-            .locator("check_box")
-            .element()
-            .expect("check_box not found");
-
-        let sub = app.subscribe().expect("subscribe");
-        chk.provider().toggle(&chk).expect("toggle failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::ValueChanged,
-                Duration::from_secs(3),
-            )
-            .expect("ValueChanged must be delivered alongside ToggleState change");
-        assert_eq!(event.kind, EventKind::ValueChanged);
-    }
-
-    /// TextChanged: setting a text field's value fires
-    /// UIA_Text_TextChangedEventId when the text provider is available.
-    /// If AccessKit's TextInput backend doesn't implement the text pattern
-    /// we fall through to ValueChanged, so accept either signal to guard
-    /// against a future AccessKit regression.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_text_changed_or_value_changed_win() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let text = find_name_field_win(&app);
-
-        let sub = app.subscribe().expect("subscribe");
-        text.provider()
-            .set_value(&text, "Event E2E Text")
-            .expect("set_value failed");
-
-        let event = sub
-            .wait_for(
-                |e| matches!(e.kind, EventKind::TextChanged | EventKind::ValueChanged),
-                Duration::from_secs(3),
-            )
-            .expect("TextChanged or ValueChanged must be delivered within 3s");
-        assert!(matches!(
-            event.kind,
-            EventKind::TextChanged | EventKind::ValueChanged
-        ));
-    }
-
-    /// Announcement: pressing the Announce button updates the live
-    /// region's value, which AccessKit's Windows bridge forwards as
-    /// `UIA_LiveRegionChangedEventId`. Our handler maps any of
-    /// `UIA_LiveRegionChangedEventId`, `UIA_NotificationEventId`, or
-    /// `UIA_SystemAlertEventId` to `EventKind::Announcement`, so we accept
-    /// all three.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "windows")]
-    fn event_announcement_win() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let btn = app
-            .locator(r#"button[name="Announce"]"#)
-            .element()
-            .expect("Announce button not found");
-
-        let sub = app.subscribe().expect("subscribe");
-        btn.provider().press(&btn).expect("Announce press failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::Announcement,
-                Duration::from_secs(3),
-            )
-            .expect("Announcement must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::Announcement);
-    }
-
-    // StructureChanged, SelectionChanged, WindowOpened / WindowClosed,
-    // WindowActivated / WindowDeactivated, MenuOpened / MenuClosed,
-    // StateChanged{Enabled|Expanded|Busy} — no end-to-end test.
-    //
-    // WindowActivated / WindowDeactivated are not emitted on Windows at
-    // all: UIA has no first-class event for them, and the design doc
-    // principle is to not model what at least two platforms don't deliver
-    // natively. (Earlier iterations of this provider inferred them from
-    // focus changes; the inference was lossy enough — false negatives on
-    // windows that open without taking focus, spurious emissions for
-    // in-app focus moves across multiple windows — to be misleading, so
-    // it was removed.)
-    //
-    // The other kinds: AccessKit's Windows bridge does not synthesize
-    // them for the widgets in our test app (single-window, no menus, no
-    // disabled/busy cycle), and some require native Win32/WPF constructs
-    // (NSMenu-style menus, live regions backed by real UIA providers)
-    // that AccessKit doesn't model. Covering them requires either a
-    // dedicated Win32 test app (future `test-apps/win32/`) or widget
-    // additions to the AccessKit app that exercise the relevant bridge
-    // code paths.
-
-    // ════════════════════════════════════════════════════════════════
-    // Event subscription (Linux-only end-to-end tests)
-    // ════════════════════════════════════════════════════════════════
-    //
-    // These tests exercise the native AT-SPI2 event subscription backend
-    // against the AccessKit test app. They are `#[ignore]` like all the
-    // other integration tests; CI runs them via `scripts/run_integ_tests.sh`
-    // (Xvfb + dbus-run-session + at-spi2-registryd).
-    //
-    // AccessKit's Linux bridge (accesskit_unix) publishes AT-SPI2 signals
-    // whenever a tree update touches the relevant properties:
-    //
-    //     AT-SPI2 signal                        → xa11y EventKind
-    //     Focus:Focus                           → FocusChanged
-    //     Object:StateChanged(focused,true)     → FocusChanged + StateChanged{Focused}
-    //     Object:StateChanged(checked,_)        → StateChanged{Checked}
-    //     Object:StateChanged(enabled,_)        → StateChanged{Enabled}
-    //     Object:PropertyChange(accessible-name) → NameChanged
-    //     Object:ValueChanged (slider/range)    → ValueChanged
-    //     Object:ValueChanged (text role)       → ValueChanged + TextChanged
-    //     Object:TextChanged                    → TextChanged
-    //     Object:ChildrenChanged                → StructureChanged
-    //     Object:SelectionChanged               → SelectionChanged
-    //     Object:Announcement                   → Announcement
-    //     Window:Create / Activate              → WindowOpened / WindowActivated
-    //     Window:Destroy / Deactivate           → WindowClosed / WindowDeactivated
-    //
-    // AT-SPI2 has no menu open/close signal, so MenuOpened/MenuClosed
-    // never fire on Linux — the design doc calls this out explicitly.
-    //
-    // Unlike the macOS tests, these tests MUST NOT catch Error::Timeout
-    // and pass silently — a hard panic on timeout is the only way to
-    // surface real regressions.
-
-    #[cfg(target_os = "linux")]
-    fn ensure_checkbox_linux(app: &App, want_on: bool) {
-        let chk = app
-            .locator("check_box")
-            .element()
-            .expect("check_box not found");
-        let is_on = chk.states.checked == Some(Toggled::On);
-        if is_on != want_on {
-            // AccessKit's AT-SPI bridge exposes checkbox toggling as the
-            // generic "click" action, not a separate "toggle" — so go
-            // through press() to match the existing action_toggle_checkbox
-            // integration test.
-            chk.provider().press(&chk).expect("press failed");
-            std::thread::sleep(std::time::Duration::from_millis(150));
-        }
-    }
-
-    // ── Subscription mechanics ──
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    fn event_try_recv_returns_none_when_idle_linux() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-
-        // Drain any signals that may have been queued during subscription setup
-        // (AT-SPI2 often replays focus/state bits to new subscribers).
-        std::thread::sleep(Duration::from_millis(300));
-        while sub.try_recv().is_some() {}
-
-        assert!(sub.try_recv().is_none());
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    fn event_recv_times_out_when_idle_linux() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-
-        std::thread::sleep(Duration::from_millis(300));
-        while sub.try_recv().is_some() {}
-
-        let r = sub.recv(Duration::from_millis(300));
-        assert!(
-            matches!(r, Err(Error::Timeout { .. })),
-            "expected Timeout, got {:?}",
-            r
-        );
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    fn event_drop_unsubscribes_cleanly_linux() {
-        let app = h::app_root();
-        {
-            let _sub = app.subscribe().expect("subscribe");
-        }
-        // Re-subscribing after drop must not hang or fail: the cancel
-        // closure removes every match rule and joins the iterator thread
-        // before returning, so the second subscribe starts clean.
-        let _sub2 = app.subscribe().expect("re-subscribe");
-    }
-
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    fn event_metadata_populated_linux() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let expected_pid = app.pid;
-
-        // Drive a deterministic event via the slider — setting a distinct
-        // value always fires Object:ValueChanged, independent of prior
-        // test state.
-        let slider = app.locator(r#"[role="slider"]"#).element().expect("slider");
-        let target_val = if slider.numeric_value == Some(42.0) {
-            17.0
-        } else {
-            42.0
-        };
-
-        let sub = app.subscribe().expect("subscribe");
-        slider
-            .provider()
-            .set_numeric_value(&slider, target_val)
-            .expect("set_numeric_value");
-
-        let event = sub
-            .wait_for(|_| true, Duration::from_secs(3))
-            .expect("at least one event must arrive within 3s");
-
-        if let Some(pid) = expected_pid {
-            assert_eq!(event.app_pid, pid, "app_pid must match subscribed app");
-        }
-        assert!(!event.app_name.is_empty(), "app_name must be populated");
-    }
-
-    // ── Per-EventKind end-to-end tests ──
-
-    /// FocusChanged: focus a non-default *enabled* button to force a real
-    /// focus move. AccessKit's AT-SPI bridge fires `StateChanged(focused,
-    /// true)` from `focus_moved()`, which we map to `FocusChanged`.
-    ///
-    /// "Enabled" matters: the test-app's initial focus is Submit; focusing
-    /// the Cancel button (disabled by default until the Terms checkbox is
-    /// ticked) produces no AT-SPI signal because AccessKit skips focus
-    /// diffs involving disabled nodes. Using the always-enabled "New"
-    /// toolbar button gives a deterministic focus transition.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    fn event_focus_changed_linux() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let target = app
-            .locator(r#"button[name="New"]"#)
-            .element()
-            .expect("New button not found");
-
-        let sub = app.subscribe().expect("subscribe");
-        target.provider().focus(&target).expect("focus failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::FocusChanged,
-                Duration::from_secs(3),
-            )
-            .expect("FocusChanged must be delivered within 3s");
-
-        assert_eq!(event.kind, EventKind::FocusChanged);
-        let tgt = event.target.expect("FocusChanged target must be populated");
-        assert_eq!(
-            tgt.role,
-            Role::Button,
-            "expected Button, got {:?}",
-            tgt.role
-        );
-    }
-
-    /// ValueChanged: setting the slider fires Object:ValueChanged on the
-    /// slider element, which the Linux backend maps to ValueChanged.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    fn event_value_changed_linux() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let slider = app
-            .locator(r#"[role="slider"]"#)
-            .element()
-            .expect("slider not found");
-
-        let sub = app.subscribe().expect("subscribe");
-        slider
-            .provider()
-            .set_numeric_value(&slider, 73.0)
-            .expect("set_numeric_value failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::ValueChanged,
-                Duration::from_secs(3),
-            )
-            .expect("ValueChanged must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::ValueChanged);
-        let tgt = event.target.expect("target");
-        assert_eq!(tgt.role, Role::Slider);
-    }
-
-    /// NameChanged: pressing Submit mutates the status label's name,
-    /// which the AccessKit bridge publishes as Object:PropertyChange
-    /// with detail="accessible-name".
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    fn event_name_changed_linux() {
-        use std::time::Duration;
-
-        // Step 1: prime status to "Please agree to terms" (checkbox off).
-        let app = h::app_root();
-        ensure_checkbox_linux(&app, false);
-        let app = h::app_root();
-        let submit = h::named(&app, "Submit");
-        submit
-            .provider()
-            .press(&submit)
-            .expect("prime press failed");
-        std::thread::sleep(Duration::from_millis(200));
-
-        // Step 2: flip checkbox on so the next Submit press drives status
-        // to "Submitted", guaranteed distinct from step 1.
-        let app = h::app_root();
-        ensure_checkbox_linux(&app, true);
-        std::thread::sleep(Duration::from_millis(200));
-
-        // Step 3: subscribe, press Submit, assert NameChanged.
-        let app = h::app_root();
-        let sub = app.subscribe().expect("subscribe");
-        let submit = h::named(&app, "Submit");
-        submit
-            .provider()
-            .press(&submit)
-            .expect("trigger press failed");
-
-        let event = sub
-            .wait_for(|e| e.kind == EventKind::NameChanged, Duration::from_secs(3))
-            .expect("NameChanged must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::NameChanged);
-
-        // Cleanup: restore checkbox to Off so later alphabetically-sorted
-        // tests (e.g. `state_checked_off_on_checkbox`,
-        // `thrash_toggle_checkbox_*`) see the same initial state. The
-        // integ suite runs `--test-threads=1` against a single long-lived
-        // test-app process, so sticky widget state carries between tests.
-        drop(sub);
-        let app = h::app_root();
-        ensure_checkbox_linux(&app, false);
-    }
-
-    // TextChanged: no end-to-end test yet on Linux.
-    //
-    // AccessKit's AT-SPI bridge only emits Object:TextChanged when the
-    // AccessKit tree's text content changes (via TextInserted / TextRemoved
-    // diffs), but the xa11y `set_value` path goes through AT-SPI's
-    // EditableText interface — which AccessKit's bridge does not implement
-    // — so the tree never observes a mutation and no signal fires. Driving
-    // real text changes on Linux requires either a GTK/Qt test harness or
-    // AccessKit test-app enhancements (e.g. exposing the text field via a
-    // bridged EditableText provider). The mapping itself is covered by
-    // the `signal_to_kinds` unit tests in xa11y-linux.
-
-    /// Announcement: pressing the "Announce" button mutates the live
-    /// region's value. AccessKit's AT-SPI bridge emits
-    /// `Object:Announcement` for `Live::Polite` nodes whenever their name
-    /// (which for `Role::Label` is derived from `value()`) changes; the
-    /// Linux backend maps that to `Announcement`.
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    fn event_announcement_linux() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let announce = app
-            .locator(r#"button[name="Announce"]"#)
-            .element()
-            .expect("Announce button not found");
-
-        let sub = app.subscribe().expect("subscribe");
-        announce.provider().press(&announce).expect("press failed");
-
-        let event = sub
-            .wait_for(
-                |e| e.kind == EventKind::Announcement,
-                Duration::from_secs(3),
-            )
-            .expect("Announcement must be delivered within 3s");
-        assert_eq!(event.kind, EventKind::Announcement);
-    }
-
-    /// StateChanged{Checked}: pressing the checkbox fires
-    /// Object:StateChanged(checked, new_value), which the Linux backend
-    /// maps to StateChanged{Checked}. AccessKit doesn't expose a distinct
-    /// "toggle" action on Linux — `press` is the canonical way to flip a
-    /// checkbox (see the pre-existing `action_toggle_checkbox` integration
-    /// test, which also uses press).
-    #[test]
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    fn event_state_changed_checked_linux() {
-        use std::time::Duration;
-        let app = h::app_root();
-        let chk = app
-            .locator("check_box")
-            .element()
-            .expect("check_box not found");
-        let was_on = chk.states.checked == Some(Toggled::On);
-
-        let sub = app.subscribe().expect("subscribe");
-        chk.provider().press(&chk).expect("press failed");
-
-        let event = sub
-            .wait_for(
-                |e| {
-                    matches!(
-                        e.kind,
-                        EventKind::StateChanged {
-                            flag: StateFlag::Checked,
-                            ..
-                        }
-                    )
-                },
-                Duration::from_secs(3),
-            )
-            .expect("StateChanged{Checked} must be delivered within 3s");
-
-        match event.kind {
-            EventKind::StateChanged {
-                flag: StateFlag::Checked,
-                value,
-            } => {
-                assert_eq!(value, !was_on, "checked flag must flip");
-            }
-            other => panic!("unexpected kind: {:?}", other),
-        }
-
-        // Restore the checkbox to its pre-test state so later tests in the
-        // same `cargo test --test-threads=1` run (e.g. `state_checked_off_
-        // on_checkbox`, `thrash_toggle_checkbox_5_times`) see the app in
-        // its expected initial state. The macOS/Windows counterparts of
-        // this test run against isolated processes in their respective
-        // harnesses, so they don't need this cleanup.
-        let app = h::app_root();
-        let chk = app
-            .locator("check_box")
-            .element()
-            .expect("check_box not found (post-test)");
-        if (chk.states.checked == Some(Toggled::On)) != was_on {
-            chk.provider()
-                .press(&chk)
-                .expect("post-test press to restore checkbox");
-        }
-    }
-
-    // Announcement, StructureChanged, SelectionChanged, Window* —
-    // AccessKit's AT-SPI bridge may or may not synthesize these for the
-    // specific widgets in the test app. They're covered by the dispatch
-    // table's unit tests (xa11y-linux::events::tests); when the bridge
-    // does emit them, the test-app-independent pathway forwards them
-    // correctly. A dedicated Linux-specific test harness (or AccessKit
-    // bridge widgets that reliably emit these signals) would be the right
-    // place to add end-to-end coverage later.
-
-    // ════════════════════════════════════════════════════════════════
-    // Helper: find text entry element
-    // ════════════════════════════════════════════════════════════════
-
-    /// Find the text entry in the app by name "Name" or by role text_field/text_area.
-    fn find_text_entry(app: &App) -> Element {
-        // Try by name first
-        let by_name = app
-            .locator(r#"[name="Name"]"#)
-            .elements()
-            .unwrap_or_default();
-        if let Some(el) = by_name.into_iter().next() {
-            return el;
-        }
-        // Fall back to text_field role
-        let fields = app.locator("text_field").elements().unwrap_or_default();
-        if let Some(el) = fields.into_iter().next() {
-            return el;
-        }
-        // Fall back to text_area role
-        let areas = app.locator("text_area").elements().unwrap_or_default();
-        areas
-            .into_iter()
-            .next()
-            .expect("Text entry not found in app")
-    }
-}
