+++ xtask/Cargo.toml
@@ -6,3 +6,4 @@ license = "MIT"
 publish = false
 
 [dependencies]
+toml = { version = "0.8", default-features = false, features = ["parse"] }
